+++ src/components/Header.tsx
@@ -1,5 +1,5 @@
 import React, { FC } from 'react';
 
-import { StyledHeader } from 'styled/header/StyledHeader';
+import { StyledHeader } from 'styled/StyledHeader';
 
 export const Header: FC = () => <StyledHeader>Test app</StyledHeader>;
