+++ src/styled/StyledFooter.tsx
